requirements.txt
flask
flask_sqlalchemy
sqlalchemy
requests
web3
py-solc-x